requirements.txt
Django==1.11.28
django-haystack==2.8.1
django-debug-toolbar==1.8
django-widget-tweaks
Whoosh==2.7.4
wagtail==2.3
wagtailcodeblock==1.15.0.0
wagtailmenus==2.8
django-el-pagination==3.1.0
markdown>=2.6.8
mysqlclient
wagtail-table-block
requests==2.18.1

nltk==3.2.4
numpy==1.13.3
pandas==0.20.3
RISparser==0.4.2
scikit-learn==0.18.2
scipy>=1.4.1
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
tflearn==0.3.2
html5lib==0.9999999
h5py
